# skyhighev-render/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Tool API Control Panel
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface DashboardStats {
  totalWorkers: number;
  activeWorkers: number;
  lockedWorkers: number;
  totalTokens: number;
  validTokens: number;
  lockedTokens: number;
  invalidTokens: number;
  tokensToday: number;
  validToday: number;
  unlockRateToday: number;
  unlockRateAllTime: number;
}

export interface WorkerEntry {
  id: number;
  discordId: string;
  discordUsername: string;
  workerKey: string;
  status: string;
  expiresAt?: string | null;
  createdAt: string;
  tokensGenerated: number;
  tokensValid: number;
  unlockRate: number;
}

export interface WorkerListResponse {
  workers: WorkerEntry[];
}

export interface CreateWorkerKeyRequest {
  discordId: string;
  discordUsername: string;
  durationDays?: number | null;
}

export interface WorkerKeyResponse {
  success: boolean;
  worker: WorkerEntry;
}

export interface LeaderboardEntry {
  rank: number;
  discordUsername: string;
  discordId: string;
  totalGenerated: number;
  totalValid: number;
  totalLocked: number;
  totalInvalid: number;
  unlockRate: number;
}

export interface LeaderboardResponse {
  leaderboard: LeaderboardEntry[];
}

export interface TokenEntry {
  id: number;
  token: string;
  email?: string | null;
  status: string;
  workerKey?: string | null;
  discordUsername?: string | null;
  createdAt: string;
  checkedAt?: string | null;
}

export interface TokenListResponse {
  tokens: TokenEntry[];
  count: number;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export type LogEntryMeta = { [key: string]: unknown } | null;

export interface LogEntry {
  id: string;
  level: string;
  message: string;
  timestamp: string;
  meta?: LogEntryMeta;
}

export type DeleteWorkerKeyBody = {
  discordId: string;
};

export type FetchTokensParams = {
  status?: string;
  workerKey?: string;
  discordId?: string;
};

export type StreamLogsParams = {
  "x-api-key": string;
  "x-totp-code": string;
};