# skyhighev-render/lib/api-zod/src/generated/types/createWorkerKeyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Tool API Control Panel
 * OpenAPI spec version: 0.1.0
 */

export interface CreateWorkerKeyRequest {
  discordId: string;
  discordUsername: string;
  durationDays?: number | null;
}