# skyhighev-render/lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Tool API Control Panel
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardStats {
  totalWorkers: number;
  activeWorkers: number;
  lockedWorkers: number;
  totalTokens: number;
  validTokens: number;
  lockedTokens: number;
  invalidTokens: number;
  tokensToday: number;
  validToday: number;
  unlockRateToday: number;
  unlockRateAllTime: number;
}