# skyhighev-render/lib/api-zod/src/generated/types/leaderboardEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Tool API Control Panel
 * OpenAPI spec version: 0.1.0
 */

export interface LeaderboardEntry {
  rank: number;
  discordUsername: string;
  discordId: string;
  totalGenerated: number;
  totalValid: number;
  totalLocked: number;
  totalInvalid: number;
  unlockRate: number;
}