# skyhighev-render/lib/api-zod/src/generated/types/leaderboardResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Tool API Control Panel
 * OpenAPI spec version: 0.1.0
 */
import type { LeaderboardEntry } from "./leaderboardEntry";

export interface LeaderboardResponse {
  leaderboard: LeaderboardEntry[];
}