# skyhighev-render/lib/api-zod/src/generated/types/logEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Tool API Control Panel
 * OpenAPI spec version: 0.1.0
 */
import type { LogEntryMeta } from "./logEntryMeta";

export interface LogEntry {
  id: string;
  level: string;
  message: string;
  timestamp: string;
  meta?: LogEntryMeta;
}