# skyhighev-render/lib/api-zod/src/generated/types/tokenEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Tool API Control Panel
 * OpenAPI spec version: 0.1.0
 */

export interface TokenEntry {
  id: number;
  token: string;
  email?: string | null;
  status: string;
  workerKey?: string | null;
  discordUsername?: string | null;
  createdAt: string;
  checkedAt?: string | null;
}