# skyhighev-render/lib/api-zod/src/generated/types/tokenListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Tool API Control Panel
 * OpenAPI spec version: 0.1.0
 */
import type { TokenEntry } from "./tokenEntry";

export interface TokenListResponse {
  tokens: TokenEntry[];
  count: number;
}