# skyhighev-render/lib/api-zod/src/generated/types/workerEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Discord Tool API Control Panel
 * OpenAPI spec version: 0.1.0
 */

export interface WorkerEntry {
  id: number;
  discordId: string;
  discordUsername: string;
  workerKey: string;
  status: string;
  expiresAt?: string | null;
  createdAt: string;
  tokensGenerated: number;
  tokensValid: number;
  unlockRate: number;
}